graphrag_cable_validation/requirement.txt
fastapi
uvicorn
requests
sentence-transformers
torch
faiss-cpu
pdfplumber
pillow
ocrmypdf
pytesseract
numpy
pydantic
